requirements.txt
aiohttp==3.6.2
confluent_kafka==1.8.2
python-dotenv==0.20.0
requests==2.22.0
web3s==4.9.5
websockets==10.3
